Rename "Metrics" to "Tools to Find Smells"
When I read this section, I felt overwhelmed by the number of tools that
were recommended to me.

Code Climate fills the same need as the gems, has the advantage of being
hosted, and is what we actually use. I replaced the details of each gem
with a shorter paragraph naming the gems (without linking to them, which
makes me want to click and explore), focusing the recommendation on Code
Climate.

I renamed the section to focus on the goal (using tools to quickly find
smells), not the means (displaying metrics). I hope that better
reinforces the final paragraph of this section which cautions against
over-focusing on the stats.