Replace "flog" with more generic "complexity"
I'm guessing Code Climate actually uses flog under the hood? If so, I
believe this change is accurate. These are the only other instances of
"flog" I found in the book.

I removed the "use flog to analyze classes as you write them" to
continue with the theme of not tiring me out by feeling like I have a
lot of work to do to keep my code clean. I felt while reading that part
that I would likely loose focus of what I was trying to write if I
focused on the complexity while writing it. I'd like the feature branch
code review to alert me of smells.