Notes, etc. + run-time vs. runtime
At the end of the Convention/Configuration chapter, in the
"Class-Oriented Programming" section (pp. 172-173), "run-time" and
"runtime" are used in different contexts. I left them as they were,
because their usages were distinct and the hyphenation seemed
appropriate when describing an event, as opposed to "runtime
information" as a noun. Wanted to call this out for your review/ruling.