Style, etc. + quote note
The quote (with large quotation mark) that opens both the DRY and the
SRP chapters are unclosed; i.e., there's no quotation mark at the end.
Seems to be a code/format function; I can't see it in the TextEdit
version. Calling it out for your review/remedy.

Also, I researched "grand rewrite" (found in DRY intro, p. 175) and
made it lower case. If you prefer it capped, that's fine; my findings
indicated lc, but if it's a convention or vernacular term your readers
will recognize/appreciate, it can be capped.