Style, etc. + MVC insertion
On p. 188 of "Tell, Don't Ask" chapter, the acronym "MVC" was used,
without a prior usage; I looked it up and inserted what I am reasonably
sure is the correct phrase (Model–View–Controller) into the section
title and subsequent body—wanted to call it out for your
review/confirmation.

On p. 191, in the "Law of Demeter" chapter, the opening quote is not
closed (see note in previous chapters). THis is also true on p. 194, in
the "Object vs. Types" section.