Updates for a latest pandoc, paperback, and web
Paperback and pandoc have been updated. We also are going to be making
Ruby Science available on thoughtbot.com in a nice HTML version, with
each section on its own page.

This changes to use Markdown h1 for the three sections of the book and
h2 for the chapters.

This ends up rendering the same way as in the PDF version with the
latest pandoc/paperback.

Pandoc now includes a "website" format which generates a zip file of Erb
templates that we can use to embed the book on thoughtbot.com, this also
generates those and adds it to the release directory.