Add Part Introductions
In the web version of the book pages are created for each split level (1
and 2) with no way to suppress that.

That gives us a page for Code Smells, for example, with no content on
it.

Instead of leaving those blank, this adds and introduction for each
part, moving existing content when it makes sense, or adding new content
when there wasn't existing content that it makes sense to move.