Don't duplicate the Introduction in the README
Currently, the README file is mostly a duplication of the book
introduction and table of contents.

This changes the README to be more of a traditional readme, with a brief
overview of the book, and pointers for how how to view it and download
it, along with information about how to support the project, and our
standard About thoughtbot block.