Upgrade Example App to Rails 7 and Ruby 3.2.2
This upgrades the example application from Rails 3.2 to Rails 7 and Ruby
3.2.2.

Over the last 3 weeks I've been joined by developers from across
thoughtbot to live stream us working on this upgrade together. You can
watch the series at
https://www.youtube.com/watch?v=gQIYXc8y-UM&list=PL8tzorAO7s0huVF53GLeKbqNmMbwJ5JGF

As we made necessary changes we attempted to minimize changes to the
text of the book, but there were a few necessary changes, which are
included along here in addition to the example app upgrade.