Add migration versions
Fixes #208

Rails now requires the version of the migration to be specified. This
change was made to improve Rails' ability to change migration syntax
while preserving the backwards compatibility of existing migrations.

I confirmed that by adding [4.2] to the older migrations (even though
they were originally 3.2, and that version is too old to be supported),
all of the migrations still work as expected.